refactor: move api calls of User Pages to UserApi.ts
