+++ package.json
@@ -14,7 +14,6 @@
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.11.1",
     "react-select": "^5.7.2",
-    "react-toastify": "^9.1.3",
     "yup": "^1.1.1"
   },
   "devDependencies": {
