+++ src/services/urls.ts
@@ -1,5 +1,3 @@
-import { updater } from "../utils/utils"
-
 export const authRoutes = {
     login: "/api/v1/yip/login/"
 }
